app/requirements.txt
asgiref==3.3.4
Django==3.2
django-cors-headers==3.7.0
djangorestframework==3.12.4
djangorestframework-gis==0.17
psycopg2-binary==2.8.6
pytz==2021.1
sqlparse==0.4.1
gunicorn==20.0.4
